Update on "[inductor] refactor wrapper benchmark code out of utils.py"
Refactor wrapper benchmark out of utils.py since
1. utils.py gets too large
2. I plan to add more code to wrapper benchmark for multi-kernel.

This is split out from https://github.com/pytorch/pytorch/pull/103469 



[ghstack-poisoned]
